requirements.txt
aiofile==3.8.8
aiofiles==23.2.1
aiopath==0.7.7
annotated-types==0.6.0
anyio==4.4.0
asyncpg==0.29.0
blinker==1.7.0
caio==0.9.17
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
ecdsa==0.18.0
fastapi==0.110.0
Flask==3.0.3
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.4
jinja2-async-environment==0.7.5
jsonschema2popo==0.12
lxml==5.2.1
MarkupSafe==2.1.5
msgspec==0.18.6
networkx==1.9
numpy==1.26.4
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
redis==5.0.4
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.36.3
Starlette-WTF==0.4.3
starlette_flash==1.1.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
WTForms==3.1.2
